pizza_broker_be/requirements.txt
uvicorn
google-adk
fastapi
httpx
fastmcp
pydantic
mysql-connector-python
CORSMiddleware